chg NULL to !

+++ push_swap-modi/stack_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   stack_utils.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: welow < welow@student.42kl.edu.my>         +#+  +:+       +#+        */
+/*   By: welow <welow@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/02 09:49:59 by utente            #+#    #+#             */
-/*   Updated: 2024/01/17 12:55:29 by welow            ###   ########.fr       */
+/*   Updated: 2024/01/23 16:21:14 by welow            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,12 +48,12 @@ void	append_node(t_stack **stack, int nbr)
 
 t_stack	*find_smallest(t_stack *stack)
 {
-	long	small;
+	int		small;
 	t_stack	*small_node;
 
 	if (!stack)
 		return (NULL);
-	small = LONG_MAX;
+	small = INT_MAX;
 	while (stack)
 	{
 		if (stack->value < small)
